+++ carbonize/api.py
@@ -1,30 +1,25 @@
 from typing import List
 
-from carbonize.calculators import FlightCalculator, TrainCalculator
+from carbonize.calculators import Flight, Train
 from carbonize.types import CarbonKg, Km, Step
 
 
 class Footprint:
-    _c_flight: FlightCalculator = FlightCalculator()
-    _c_train: TrainCalculator = TrainCalculator()
-
     def __init__(self):
         self.steps: List[Step] = []
 
-    def add_flight(self, *, a: str, b: str, two_way: bool = False) -> None:
-        step = self._c_flight.calculate(a=a, b=b)
-        self.steps.append(step)
+    def add_flight(self, *, a: str, b: str, two_way: bool = False, aircaft: str = Flight.DEFAULT_AIRCRAFT) -> None:
+        self.steps.append(Flight(a=a, b=b).get_step())
 
         if two_way:
-            self.add_flight(a=b, b=a)
+            self.add_flight(a=b, b=a, aircaft=aircaft)
 
     def add_train(self, *, distance: Km, two_way: bool = False) -> None:
-        step = self._c_train.calculate(distance=distance)
-        self.steps.append(step)
+        self.steps.append(Train(distance=distance).get_step())
 
         if two_way:
             self.add_train(distance=distance)
 
     @property
     def emissions(self) -> CarbonKg:
-        return CarbonKg(sum([step.emissions for step in self.steps]))
+        return CarbonKg(sum([step.co2e for step in self.steps]))