+++ carbonize/calculators/trains.py
@@ -3,7 +3,7 @@ from carbonize.types import CarbonKg, Km, Step
 from .base import Calculator
 
 
-class TrainCalculator(Calculator):
+class Train(Calculator):
     """A carbon emissions calculator for railway transport.
     Based on UK GOV carbon conversion factors (kg/km):
     https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2019
@@ -11,5 +11,11 @@ class TrainCalculator(Calculator):
 
     CO2E_PER_KM = 0.04115
 
-    def calculate(self, *, distance: Km) -> Step:
-        return Step(distance, CarbonKg(distance * self.CO2E_PER_KM))
+    def __init__(self, *, distance: Km) -> None:
+        self.distance = distance
+
+    def __repr__(self) -> str:
+        return f"Train({self.distance} km)"
+
+    def get_step(self) -> Step:
+        return Step(CarbonKg(self.distance * self.CO2E_PER_KM), self.distance, None, self)