+++ carbonize/calculators/webinars.py
@@ -1,1 +1,9 @@
-# https://blog.zoom.us/how-video-meetings-are-helping-reduce-environmental-impact-infographic/
+from .base import Calculator
+
+
+class Webinar(Calculator):
+    """A carbon emissions calculator for online webinars.
+    https://blog.zoom.us/how-video-meetings-are-helping-reduce-environmental-impact-infographic/
+    """
+
+    pass