+++ CHANGELOG.md
@@ -5,8 +5,23 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-### Added
-- Initial wallets, based on `lnbits` wallets.
+## [0.1.0] - 2022-xxx
 
-[unreleased]: https://github.com/python-ln/lnurl/compare/0.0.1...HEAD
-[0.0.1]: https://github.com/python-ln/lnurl/releases/tag/0.0.1
+### Added
+- General API, introducing the Footprint class and Steps.
+- New calculator for trains: `Train`.
+
+### Changed
+- Underlying data is stored as Pickle files instead of CSV files.
+- Save typed data to reduce processing when catalogs are loaded.
+- `AviationCalculator` is now `Flight`.
+
+## [0.0.1] - 2018-04-13
+
+### Added
+- First calculator: `AviationCalculator`.
+- Initial data.
+
+[unreleased]: https://github.com/eillarra/carbonize/compare/0.1.0...HEAD
+[0.1.0]: https://github.com/eillarra/carbonize/releases/tag/0.1.0
+[0.0.1]: https://github.com/eillarra/carbonize/releases/tag/0.0.1